requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
psycopg[binary]>=3.1.0
psycopg_pool
requests==2.32.4
python-dotenv==1.0.0
gunicorn==20.1.0
python-dateutil==2.8.2
sentry-sdk[flask]>=1.40.0
redis>=4.0.0
